requirements.txt
matplotlib >= 1.5.1
mpi4py >= 2.0.0
numpy >= 1.11.0
beautifulsoup4
pillow


